httpclient - upgrade to 4.3.6
Needed for having GN being able to request TLSv1.1 / 1.2 only HTTPS
servers (as used in geOrchestra dockerized envs).

The main issue with httpclient is that it discards the value of the
system variable https.protocols. v4.3.6 of the lib activates TLS (and
also deactivates SSLv3).

References:

- https://www.apache.org/dist/httpcomponents/httpclient/RELEASE_NOTES-4.3.x.txt
- https://github.com/camptocamp/georchestra-pigma-configuration/issues/328

Tests:
On a vanilla GeoNetwork 3.0.x version:
- compilation OK, utests skipped
- Runtime tests OK

Since this is a minor upgrade (4.3.2 to 4.3.6), this change is
considered safe.